Move JUnit stuff into its own package
This may become its own module in the future
