+++ src/main/java/io/github/arthurc/autotest/junit/jupiter/TestPlanLifecycleTestExecutionListener.java
@@ -1,7 +1,7 @@
 /**
  * SPDX-License-Identifier: MIT
  */
-package io.github.arthurc.autotest.testplan.junit.jupiter;
+package io.github.arthurc.autotest.junit.jupiter;
 
 import io.github.arthurc.autotest.lifecycle.Lifecycle;
 import io.github.arthurc.autotest.testplan.TestPlanLifecycle;
