+++ src/main/resources/META-INF/services/org.junit.platform.launcher.TestExecutionListener
@@ -1,2 +1,2 @@
-io.github.arthurc.autotest.testplan.junit.jupiter.TestPlanLifecycleTestExecutionListener
-io.github.arthurc.autotest.testexecution.junit.jupiter.TestExecutionLifecycleTestExecutionListener
+io.github.arthurc.autotest.junit.jupiter.TestPlanLifecycleTestExecutionListener
+io.github.arthurc.autotest.junit.jupiter.TestExecutionLifecycleTestExecutionListener
