+++ src/test/java/io/github/arthurc/autotest/junit/jupiter/TestExecutionLifecycleTestExecutionListenerTest.java
@@ -1,4 +1,4 @@
-package io.github.arthurc.autotest.testexecution.junit.jupiter;
+package io.github.arthurc.autotest.junit.jupiter;
 
 import io.github.arthurc.autotest.test.utils.EventCollector;
 import io.github.arthurc.autotest.testexecution.TestExecutionLifecycle;
