+++ src/test/java/io/github/arthurc/autotest/junit/jupiter/TestPlanLifecycleTestExecutionListenerTest.java
@@ -1,4 +1,4 @@
-package io.github.arthurc.autotest.testplan.junit.jupiter;
+package io.github.arthurc.autotest.junit.jupiter;
 
 import io.github.arthurc.autotest.AutotestTest;
 import io.github.arthurc.autotest.example.ExampleApp;
