+++ src/test/resources/junit-platform.properties
@@ -1,4 +1,4 @@
 junit.jupiter.displayname.generator.default=\
     org.junit.jupiter.api.DisplayNameGenerator$ReplaceUnderscores
 junit.platform.execution.listeners.deactivate=\
-  io.github.arthurc.autotest.testplan.junit.jupiter.TestPlanLifecycleTestExecutionListener
+  io.github.arthurc.autotest.junit.jupiter.TestPlanLifecycleTestExecutionListener
